+++ src/main/java/http/request/RequestMethod.java
@@ -1,4 +1,4 @@
-package dto.request;
+package http.request;
 
 import exceptions.InvalidRequestMethodException;
