+++ src/main/java/http/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package dto;
+package http.response;
 
 
 import utils.FileIoUtils;
